fix: change totalPrice type

+++ src/main/java/org/kys/bnmo/model/Transaction.java
@@ -21,7 +21,7 @@ public class Transaction implements Serializable {
     private int quantity;
 
     @Getter
-    private int totalPrice;
+    private Double totalPrice;
 
     @Getter
     private Date date;
